backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers
psycopg[binary]>=3.1
python-dotenv>=1.0

pytest>=8.0
pytest-django>=4.8
coverage>=7.0
openpyxl>=3.1
reportlab>=4.0